Moved invoices rpc server to rpcservers folder
