+++ log.go
@@ -21,7 +21,6 @@ import (
 	"github.com/lightningnetwork/lnd/healthcheck"
 	"github.com/lightningnetwork/lnd/htlcswitch"
 	"github.com/lightningnetwork/lnd/invoices"
-	"github.com/lightningnetwork/lnd/lnrpc/invoicesrpc"
 	"github.com/lightningnetwork/lnd/lnrpc/neutrinorpc"
 	"github.com/lightningnetwork/lnd/lnrpc/peersrpc"
 	"github.com/lightningnetwork/lnd/lnrpc/routerrpc"
@@ -42,6 +41,7 @@ import (
 	autopilotrpc "github.com/lightningnetwork/lnd/rpcservers/autopilot"
 	chainrpc "github.com/lightningnetwork/lnd/rpcservers/chain"
 	devrpc "github.com/lightningnetwork/lnd/rpcservers/dev"
+	invoicesrpc "github.com/lightningnetwork/lnd/rpcservers/invoices"
 	"github.com/lightningnetwork/lnd/signal"
 	"github.com/lightningnetwork/lnd/sweep"
 	"github.com/lightningnetwork/lnd/tor"
