+++ rpcservers/invoices/addinvoice.go
@@ -1,4 +1,4 @@
-package invoicesrpc
+package invoices
 
 import (
 	"bytes"
@@ -191,7 +191,6 @@ func (d *AddInvoiceData) mppPaymentHashAndPreimage() (*lntypes.Preimage, lntypes
 	)
 
 	switch {
-
 	// Only either preimage or hash can be set.
 	case d.Preimage != nil && d.Hash != nil:
 		return nil, lntypes.Hash{},
@@ -321,7 +320,7 @@ func AddInvoice(ctx context.Context, cfg *AddInvoiceConfig,
 	// If not, use the memo field as the payment request description.
 	if len(invoice.DescriptionHash) > 0 {
 		var descHash [32]byte
-		copy(descHash[:], invoice.DescriptionHash[:])
+		copy(descHash[:], invoice.DescriptionHash)
 		options = append(options, zpay32.DescriptionHash(descHash))
 	} else {
 		// Use the memo field as the description. If this is not set
