+++ rpcservers/invoices/config_default.go
@@ -1,7 +1,7 @@
 //go:build !invoicesrpc
 // +build !invoicesrpc
 
-package invoicesrpc
+package invoices
 
 // Config is empty for non-invoicesrpc builds.
 type Config struct{}
