+++ subrpcserver_config.go
@@ -13,7 +13,6 @@ import (
 	"github.com/lightningnetwork/lnd/htlcswitch"
 	"github.com/lightningnetwork/lnd/invoices"
 	"github.com/lightningnetwork/lnd/lncfg"
-	"github.com/lightningnetwork/lnd/lnrpc/invoicesrpc"
 	"github.com/lightningnetwork/lnd/lnrpc/neutrinorpc"
 	"github.com/lightningnetwork/lnd/lnrpc/peersrpc"
 	"github.com/lightningnetwork/lnd/lnrpc/routerrpc"
@@ -28,6 +27,7 @@ import (
 	autopilotrpc "github.com/lightningnetwork/lnd/rpcservers/autopilot"
 	chainrpc "github.com/lightningnetwork/lnd/rpcservers/chain"
 	devrpc "github.com/lightningnetwork/lnd/rpcservers/dev"
+	invoicesrpc "github.com/lightningnetwork/lnd/rpcservers/invoices"
 	"github.com/lightningnetwork/lnd/sweep"
 	"github.com/lightningnetwork/lnd/watchtower"
 	"github.com/lightningnetwork/lnd/watchtower/wtclient"
